Move struct definitions in llama.cpp to llama.h
